feat(lambda): replace all updateDeviceShadow feature into getDeviceShadow
