+++ aws-backend/infrastructure/iam/iam-parking-stack.yaml
@@ -107,10 +107,6 @@ Resources:
                   - sqs:DeleteMessage
                   - sqs:GetQueueAttributes
                 Resource: "*"
-              - Effect: Allow
-                Action:
-                  - secretsmanager:GetSecretValue
-                Resource: !Sub arn:aws:secretsmanager:${AWS::Region}:${AWS::AccountId}:secret:parkingdb-db-credentials-${EnvironmentKey}-*
 
   ## Role for GetDeviceShadow Lambda
   GetDeviceShadowLambdaRole:
@@ -137,6 +133,10 @@ Resources:
                 Action:
                   - iot:GetThingShadow
                 Resource: "*"
+              - Effect: Allow
+                Action:
+                  - secretsmanager:GetSecretValue
+                Resource: !Sub arn:aws:secretsmanager:${AWS::Region}:${AWS::AccountId}:secret:parkingdb-db-credentials-${EnvironmentKey}-*
 
   ## Role for Shadow Delta Handler Lambda
   ShadowDeltaHandlerRole:
